+++ examples/read_texts.rs
@@ -12,11 +12,11 @@ fn main() {
     let mut buf = Vec::new();
 
     loop {
-        match reader.read_event(&mut buf) {
+        match reader.read_event_into(&mut buf) {
             Ok(Event::Start(ref e)) if e.name().as_ref() == b"tag2" => {
                 txt.push(
                     reader
-                        .read_text(b"tag2", &mut Vec::new())
+                        .read_text_into(b"tag2", &mut Vec::new())
                         .expect("Cannot decode text value"),
                 );
                 println!("{:?}", txt);